requirements.txt
django==3.2.16
djangorestframework==3.14.0
pillow==10.0.1
psycopg2-binary==2.9.9
django-environ==0.11.2
coverage==7.3.2
django-filter==23.3
coreschema==0.0.4
drf-yasg==1.21.7